refactor libs/cloudformation, adds more ava tests
