+++ package.json
@@ -8,14 +8,14 @@
     "clean": "rm -rf dist/",
     "compile": "babel -s true --source-maps inline -d dist/ src/ --ignore *.AppleDouble*",
     "link:watch": "nodemon -w src -x npm -- run lint-link",
-    "format": "prettier-eslint --write src/**/*.js | snazzy",
-    "lint": "npm run format",
+    "format": "prettier-eslint --write src/**/*.js",
+    "lint": "eslint src/**/*.js",
     "lint-link": "npm run lint && npm link",
     "lint:watch": "nodemon -w src -x npm -- run lint",
     "prepublish": "npm run clean && npm run compile",
     "pre-commit": "npm run lint",
     "coverage-report": "nyc report --check-coverage --branches 100 --functions 100 --lines 100 --statements 100 --reporter=lcov",
-    "test": "nyc ava src/**/*.spec.js && npm run coverage-report"
+    "test": "npm run lint && nyc ava src/**/*.spec.js && npm run coverage-report"
   },
   "engines": {
     "node": ">=6"
@@ -65,6 +65,7 @@
     "eslint": "^3.14.1",
     "eslint-config-semistandard": "^7.0.0",
     "eslint-config-standard": "^6.2.1",
+    "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-promise": "^3.4.0",
     "eslint-plugin-standard": "^2.0.1",
     "nyc": "^10.1.2",
