+++ src/commands/describe.js
@@ -5,9 +5,9 @@ import loadConfig from '../config';
 import { error, log, title } from '../logger';
 import {
   getStackOutputs,
-  getStackResources,
-  templateStackName
-} from '../libs/cloudformation';
+  getStackResources
+} from '../libs/aws/cfn-get-stack-info-helpers';
+import { templateStackName } from '../factories/cloudformation';
 
 export function run (argv) {
   const { APP_NAME } = loadConfig();
